requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
debugpy==1.5.1
decorator==5.1.0
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.3
djangorestframework==3.12.4
drf-extensions==0.7.1
entrypoints==0.3
gunicorn==20.1.0
idna==3.3
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
